yaml rollback, message in crash tip fixed and exceptions moved
