+++ tests/unit_tests/detection_dataset_test.py
@@ -6,7 +6,7 @@ from torch.utils.data import DataLoader
 from super_gradients.training.dataloaders import coco2017_train_yolo_nas, get_data_loader
 from super_gradients.training.datasets import COCODetectionDataset
 from super_gradients.training.datasets.data_formats.default_formats import LABEL_CXCYWH
-from super_gradients.training.exceptions.dataset_exceptions import DatasetValidationException, ParameterMismatchException
+from super_gradients.common.exceptions.dataset_exceptions import DatasetValidationException, ParameterMismatchException
 from super_gradients.training.transforms import DetectionMosaic, DetectionTargetsFormatTransform, DetectionPaddedRescale
