+++ tests/unit_tests/train_with_precise_bn_test.py
@@ -27,7 +27,7 @@ class TrainWithPreciseBNTest(unittest.TestCase):
             "optimizer_params": {"weight_decay": 1e-4, "momentum": 0.9},
             "train_metrics_list": [Accuracy(), Top5()],
             "valid_metrics_list": [Accuracy(), Top5()],
-            "metric_to_watch": "Accuracy",
+            "metric_to_watch": "Accuracy345",
             "greater_metric_to_watch_is_better": True,
             "precise_bn": True,
             "precise_bn_batch_size": 100,
